// src/openapi/model/addDocOfAutolivCommand.ts
/**
 * ICNC.ERP.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AddDocOfAutolivCommand { 
    product_id?: string;
    type?: string | null;
    file_name?: string | null;
    bucket?: string | null;
}